feat: collect signatures
