+++ packages/did-tools/src/types/method-signature.ts
@@ -1,0 +1,5 @@
+export interface MethodSignature {
+  name: string;
+  paramType: string[];
+  returnType: string;
+}
